+++ sdk/resourcemanager/azure-resourcemanager-cdn/src/main/java/com/azure/resourcemanager/cdn/fluent/models/EdgeNodeInner.java
@@ -7,20 +7,15 @@ package com.azure.resourcemanager.cdn.fluent.models;
 import com.azure.core.annotation.Fluent;
 import com.azure.core.management.ProxyResource;
 import com.azure.core.management.SystemData;
-import com.azure.core.util.logging.ClientLogger;
 import com.azure.resourcemanager.cdn.models.IpAddressGroup;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 
 /** Edgenode is a global Point of Presence (POP) location used to deliver CDN content to end users. */
 @Fluent
 public final class EdgeNodeInner extends ProxyResource {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(EdgeNodeInner.class);
-
     /*
-     * The JSON object that contains the properties required to create an
-     * edgenode.
+     * The JSON object that contains the properties required to create an edgenode.
      */
     @JsonProperty(value = "properties")
     private EdgeNodeProperties innerProperties;