+++ sdk/resourcemanager/azure-resourcemanager-cdn/src/main/java/com/azure/resourcemanager/cdn/models/CertificateSource.java
@@ -8,7 +8,7 @@ import com.azure.core.util.ExpandableStringEnum;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import java.util.Collection;
 
-/** Defines values for CertificateSource. */
+/** Defines the source of the SSL certificate. */
 public final class CertificateSource extends ExpandableStringEnum<CertificateSource> {
     /** Static value AzureKeyVault for CertificateSource. */
     public static final CertificateSource AZURE_KEY_VAULT = fromString("AzureKeyVault");
@@ -27,7 +27,11 @@ public final class CertificateSource extends ExpandableStringEnum<CertificateSou
         return fromString(name, CertificateSource.class);
     }
 
-    /** @return known CertificateSource values. */
+    /**
+     * Gets known CertificateSource values.
+     *
+     * @return known CertificateSource values.
+     */
     public static Collection<CertificateSource> values() {
         return values(CertificateSource.class);
     }