+++ sdk/resourcemanager/azure-resourcemanager-cosmos/src/main/java/com/azure/resourcemanager/cosmos/models/AutoUpgradePolicyResource.java
@@ -11,8 +11,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @Fluent
 public final class AutoUpgradePolicyResource {
     /*
-     * Represents throughput policy which service must adhere to for
-     * auto-upgrade
+     * Represents throughput policy which service must adhere to for auto-upgrade
      */
     @JsonProperty(value = "throughputPolicy")
     private ThroughputPolicyResource throughputPolicy;