+++ sdk/storage/azure-storage-file-share/src/test/java/com/azure/storage/file/share/FileAPITests.groovy
@@ -417,7 +417,7 @@ class FileAPITests extends APISpec {
 
         data.defaultBytes == stream.toByteArray()
     }
-
+    
     def "Download all null"() {
         given:
         primaryFileClient.create(data.defaultDataSizeLong)