+++ sdk/confidentialledger/azure-security-confidentialledger/CHANGELOG.md
@@ -10,6 +10,14 @@
 
 ### Other Changes
 
+## 1.0.1 (2022-08-17)
+
+### Other Changes
+
+#### Dependency Updates
+- Upgraded `azure-core` dependency to `1.31.0`.
+- Upgraded `azure-core-http-netty` dependency to `1.12.4`.
+
 ## 1.0.0 (2022-07-18)
 - GA ("Generally Available") release for 2022-05-13 service version. 
 ### New Features